Refactor code to make it more structured
