+++ src/main/java/Main.java
@@ -1,0 +1,6 @@
+import impl.Danny;
+public class Main {
+    public static void main(String[] args){
+        new Danny();
+    }
+}
