+++ src/main/java/exceptions/InvalidCommandException.java
@@ -1,0 +1,7 @@
+package exceptions;
+
+public class InvalidCommandException extends Exception{
+    public InvalidCommandException(String message){
+        super(message);
+    }
+}
