+++ src/main/java/impl/interfaces/Events.java
@@ -1,3 +1,5 @@
+package impl.interfaces;
+
 public class Events extends Task{
     protected String by;
     protected String from;
