+++ src/main/java/impl/interfaces/Task.java
@@ -1,3 +1,5 @@
+package impl.interfaces;
+
 public class Task {
     protected String description;
     protected boolean isDone;
