+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
-Hello! I'm Danny
+Hello! I'm Danny!
 What can I do for you?
 ____________________________________________________________
 ____________________________________________________________
